more changes

+++ hello.py
@@ -5,7 +5,7 @@ from datetime import datetime
 from datetime import date
 from werkzeug.security import generate_password_hash, check_password_hash
 from flask_login import UserMixin, login_user, LoginManager, login_required, logout_user, current_user
-from webforms import LoginForm, PostForm, UserForm, PasswordForm, NamerForm
+from webforms import LoginForm, PostForm, UserForm, PasswordForm, NamerForm, SearchForm
 
 # Create a Flask Instance
 app = Flask(__name__)
@@ -28,6 +28,26 @@ login_manager.login_view = 'login'
 @login_manager.user_loader
 def load_user(user_id):
 	return Users.query.get(int(user_id))
+
+# Pass Stuff to Navbar
+@app.context_processor
+def base():
+	form = SearchForm()
+	return dict(form=form)
+
+# Create Search Function
+@app.route('/search', methods=["POST"])
+def search():
+	form = SearchForm()
+	posts = Posts.query
+	if form.validate_on_submit():
+		# Get data from submitted form
+		post.searched = form.searched.data
+		# Query the database
+		posts = posts.filter(Posts.content.like('%' + post.searched + '%'))
+		posts = posts.order_by(Posts.title).all()
+		return render_template("search.html", form=form, searched=post.searched, posts=posts)
+
 
 # Create Login Page
 @app.route('/login', methods=['GET', 'POST'])
@@ -97,9 +117,10 @@ def post(id):
 def edit_post(id):
 	post = Posts.query.get_or_404(id)
 	form = PostForm()
+
 	if form.validate_on_submit():
 		post.title=form.title.data
-		post.author = form.author.data
+		# post.author = form.author.data
 		post.slug = form.slug.data
 		post.content = form.content.data
 		# Update database
@@ -107,28 +128,43 @@ def edit_post(id):
 		db.session.commit()
 		flash("Post has been updated!")
 		return redirect(url_for('post', id=post.id))
-	form.title.data = post.title
-	form.author.data = post.author
-	form.slug.data = post.slug
-	form.content.data = post.content
-	return render_template('edit_post.html', form=form)	
+
+	if current_user.id == post.poster_id:
+		form.title.data = post.title
+		# form.author.data = post.author
+		form.slug.data = post.slug
+		form.content.data = post.content
+		return render_template('edit_post.html', form=form)
+	else:
+		flash("You are not authorised to Edit this Post!")
+		posts = Posts.query.order_by(Posts.date_posted)
+		return render_template("posts.html", posts=posts)		
 
 # Delete Posts
 @app.route('/post/delete/<int:id>')
 @login_required
 def delete_post(id):
 	post_to_delete = Posts.query.get_or_404(id)
-	try:
-		db.session.delete(post_to_delete)
-		db.session.commit()
+	id = current_user.id
+	if id == post_to_delete.poster.id:
+		try:
+			db.session.delete(post_to_delete)
+			db.session.commit()
+			# Return a message
+			flash("Blog Post Was Deleted!")
+			posts = Posts.query.order_by(Posts.date_posted)
+			return render_template("posts.html", posts=posts)
+
+		except:
+			flash("Sorry! There was an error deleting. Please try again.")
+			posts = Posts.query.order_by(Posts.date_posted)
+			return render_template("posts.html", posts=posts)		
+
+	else:
 		# Return a message
-		flash("Blog Post Was Deleted!")
+		flash("You aren't authorised to delete that post!")
 		posts = Posts.query.order_by(Posts.date_posted)
 		return render_template("posts.html", posts=posts)
-	except:
-		flash("Sorry! There was an error deleting. Please try again.")
-		posts = Posts.query.order_by(Posts.date_posted)
-		return render_template("posts.html", posts=posts)		
 
 # Add Post Page
 @app.route('/add-post', methods=['GET', 'POST'])
@@ -136,11 +172,12 @@ def delete_post(id):
 def add_post():
 	form = PostForm()
 	if form.validate_on_submit():
-		post = Posts(title=form.title.data, content=form.content.data, author=form.author.data, slug=form.slug.data)
+		poster = current_user.id
+		post = Posts(title=form.title.data, content=form.content.data, poster_id=poster, slug=form.slug.data)
 		# Clear the form
 		form.title.data = ''
 		form.content.data = ''
-		form.author.data = ''
+		# form.author.data = ''
 		form.slug.data = ''
 		# Add post data to database
 		db.session.add(post)
@@ -281,9 +318,11 @@ class Posts(db.Model):
 	id = db.Column(db.Integer, primary_key=True)
 	title = db.Column(db.String(255))
 	content = db.Column(db.Text)
-	author = db.Column(db.String(255))
+	# author = db.Column(db.String(255))
 	date_posted = db.Column(db.DateTime, default=datetime.utcnow)
 	slug = db.Column(db.String(255))
+	# Foreign Key To Link Users (refer to primary key of the user)
+	poster_id = db.Column(db.Integer, db.ForeignKey('users.id'))
 
 # Create Model
 class Users(db.Model, UserMixin):
@@ -295,6 +334,8 @@ class Users(db.Model, UserMixin):
 	date_added = db.Column(db.DateTime, default=datetime.utcnow)
 	# Do some password stuff!
 	password_hash = db.Column(db.String(128))
+	# User can have many posts
+	posts = db.relationship('Posts', backref='poster')
 
 	@property
 	def password(self):
